feat: factorisation composants
